refactor: code structure.
