+++ src/app/posts/[slug]/page.tsx
@@ -1,6 +1,6 @@
-import { getPostData } from '@/lib/posts';
+import { getPostData } from '@/services/posts';
 import { PostData } from '@/types';
-import PostLayout from '@/components/PostLayout';
+import PostLayout from '@/components/layout/PostLayout';
 
 interface PostPageProps {
   params: {
