+++ src/services/getConfig.ts
